src/agent-host/pyproject.toml
[tool.poetry]
name = "agent-host"
version = "0.1.0"
description = "Agent Host - BFF service providing chat interface for MCP Tools Provider"
authors = ["Bruno van de Werve <bvandewe@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
neuroglia-python = { version = "^0.7.8", extras = ["mongodb", "eventstore"] }
redis = "^7.1.0"
authlib = "^1.3.0"
itsdangerous = "^2.2.0"
pyjwt = { version = "^2.8.0", extras = ["crypto"] }
httpx = "^0.27.0"
multipledispatch = "^1.0.0"
pyyaml = "^6.0.1"
websockets = "^15.0.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
ruff = ">=0.1.0"
pytest-asyncio = "^1.3.0"
pytest-playwright = "^0.7.0"

[tool.poetry.group.e2e]
optional = true

[tool.poetry.group.e2e.dependencies]
pytest-playwright = "^0.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 200
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true
known-first-party = ["domain", "application", "api", "infrastructure", "integration", "ui", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "strict"
addopts = "-v"
markers = [
    "e2e: marks tests as end-to-end tests (deselect with '-m \"not e2e\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
